Add LGTM config file and explicitly set Python version
For whatever reason, when LGTM extracts code from this project, its
interpretConfig tool (a Java JAR, the source for which I can't find)
generates an effective config that tells the analysis to use Python 2
despite all of our tooling having been updated to drop that version.

It would be nice to get rid of the false positive alerts we have about
py2 stuff, so in the absence of any tooling (that I can find) to track
down why LGTM infers the wrong Python version... Let's just set it.