irc: improve behavior of ModeParser.parse method
Several improvements of the ModeParser.parse method, which was renamed
from ModeParser.parse_modestring (renamed to prevent name conflict with
sopel.irc.modes.parse_modestring function).

The method ModeParser.get_mode_info doesn't try to parse privileges
anymore, which make it more type-safe as the letter is now str instead
of Optional[str]. This makes the logic in ModeParser.parse a little
different in a good way.

Overall, a better type definition helped built a better interface with
less potential for issues later. Functionnal tests for ModeParser.parse
are kept the same, as nothing has change on this side.

Thanks to mal who caught the type inconsistency in the first place!