reddit: fix matching twice for reddit.com without www.
Addresses a regression introduced by #2202, which added `reddit.com` to
the `short_post_url` pattern and made it possible for a post or comment
URL to match BOTH that and `post_or_comment_url`.

Since URLs match individually, and not as part of the line, all we need
to make sure a shortlink really is one is to ensure that the link ends
after the submission ID and optional trailing slash.