cli: fix legacy parser options for Py3.9.8
The hack adds a prefix to the "dest" of the legacy options added to the
main parser. It works because these options are not actually used by the
legacy subparser: they exists only to trick the help message into
something usable by end-users. It's rather ugly, but at this point we
already know that the "legacy" subparser trick is ugly.

We are only fixing the issue introduced by Python 3.9.8 that "fixed"
something in argparse that broke our use-case. Let's not delve into
that for now.