docs: fix not-found references in plugins.rules
Given the choice between updating the cross-references (this change) or
telling Sphinx to include `:undoc-members:` and/or `:inherited-members:`
that might not have docstrings and would thus require further work (in
the form of config changes, another new extension, and/or making the
code copy over the parent class's method docstring)..........

I chose to update the cross-references. It's the least amount of work,
and yields docs with minimal duplication between classes. We don't need
to repeat the docstring for `from_callable` everywhere.