tools: move Identifier to `tools.identifiers`
And fix the default casemapping to follow the RFC it is supposed to
implement (RFC 1459): only A-Z letters should be lowercased.

The lowercase function is now part of a set of public functions, each
implement a different "lower" rule:

* ascii: implement the "ascii" CASEMAPPING, i.e. A-Z only
* rfc1459: implement the "rfc1459" CASEMAPPING, i.e. A-Z + []\~
* rfc1459-strict: implement the same RFC but correctly, i.e. no
  replacement for ~ (it shouldn't be)

The `Identifier` class now use the rfc1459 function, instead of using
str.lower.

In order to propagate this change correctly, I updated `sopel.db` as
well, because str.lower() is not equivalent to Identifier.lower()
anymore.

Note that this doesn't support unicode Identifier, which would use
str.lower directly; the "PRECIS" version of the IRC v3 spec is not ready
yet, so it is not implemented for the moment. Also, it would requires to
manage not only CASEMAPPING, but also UTF8MAPPING, and that's for
another day/feature!

See also https://modern.ircdocs.horse/index.html#casemapping-parameter
and any links from that document.

Co-authored-by: dgw <dgw@technobabbl.es>