bot, modules: use bot.make_identifier
Instead of instantiating `tools.Identifier` directly, built-in plugins
are now using `bot.make_identifier(nick)` to get an Identifier for
nicks and channel. This method comes from `sopel.irc.AbstractBot`, and
can be used as an Identifier factory: take a string as input, return an
Identifier.

In the future, this factory will be able to take advantage of the
configuration and the ISUPPORT to select the appropriate CASEMAPPING
function.

The next steps are:

* to use that factory outside of plugins (i.e. triggers and targets)
* to implement a different CASEMAPPING depending on context
* add documentation about "how to get an Identifier"