config: core.nick is now a str, not Identifier
The core configuration "nick" option was an Identifier: it was
convenient to always have an Identifier for the configured nick.
However, it doesn't make a lot of sense: the nick is used in various
place as it is (no lowercase), and where it needs to be an Identifier it
is always to compare to to another one.

As a result:

* type(core.nick) is now str
* when comparing the config value to an identifier, use
  bot.make_identifier