db, test: do not create nick ID by default
Having `get_nick_id()` default to `create=True` may cause unexpected
behavior if invalid nicknames are passed to it via plugins' use of the
getter and setter methods.

We are taking this opportunity to fix the API design itself, with no
period of deprecation and warnings, since this particular method is
mostly used only by the database object itself, and isn't really
intended for direct use by plugins.

I've also added some more assertions and error-case checking to some of
the related tests for `sopel.db`.