docs: fix TypeVar autodoc warning
It appears that Sphinx's autodoc plugin doesn't understand TypeVar
properly: it doesn't generate documentation, and it doesn't
understand it when used in a function's signature.

To fix that warning for now, a fake class declaration is written
in the documentation.

I added a TODO for when Sphinx fixes that problem.