adminchannel: check kicks against bot's *current* nick
I traced use of the equivalent to `bot.config.core.nick` all the way
back to the 3.x (Phenny) days. Couldn't see a reason why this check
should use the configuration value. It just always has, probably because
`bot.nick` hasn't always meant what it means in the Sopel of today.

Since the only logical reason to check the nick is to prevent someone
from making the bot kick itself, that's what we'll make sure it does.
And since the bot's nick can theoretically change at runtime, the check
should use the bot's *current* nick, not the config value.