coretasks: additional handling of forced nick-change
It's a good idea to make sure the bot always knows its current nick,
even if that value doesn't match what it wants (from config).

If `echo-message` is available, for example, the server will send lines
using the bot's *current* nick, not the configured one. Sopel simulates
this if the capability isn't enabled. We want e.g. the `plugin.echo`
decorator to still work regardless.