xkcd: account for Bing search returning mobile URLs
Interestingly, adding '-site:m.xkcd.com' made Bing return NO results,
not the normal (non-mobile) link. That's why this patch modifies the
regex to validate the result instead.

I'm not sure what's up with the optional protocol prefix, but didn't
want to touch it in a patch intended for stable branch.