tools: fix sendable message excess truncated when splitting by spaces
This is a stupid little tweak that is obvious in hindsight, since the
code path above it to do the same when there is no space to split on
already had this logic down right. I have plugged this gap in our test
suite, as well, with a fun copypasta.