plugins.handlers: entrypoint starts with module version if available
The other way to do this would be what was there before this change,
plus an added `if version is None:` that then calls on `super()`. I'm
just not convinced that that would be better, and it's more LoC. Plus,
in theory, an entrypoint's module could have its own separate version
number that differs from the package's version, and this order of
operations allows showing the more specific one if it's set.