ci: add Python 3.10 and update pytest
This step is fairly easy:

* add `"3.10"` in `ci.yml` (the quotes are important because YAML)
* upgrade pytest to 7.1
* add an option to pytest

The annoying problem with pytest 7.x is the nose plugin: activated by
default, it is used as a compatiblity plugin with old codebase that used
the nosetests testing framework and switched to pytest. However, for
some reason, it automatically consider the `setup` (and `teardown`)
function as a setup hook for test, making it into a pytest fixture, for
which the first argument is the module itself.

The solution is to disable the nose plugin in our configuration.