plugins: replace imp by importlib to find plugins
Before, we used a `setup.py` file and the "old style" Python packaging
with setuptools. And we didn't used Python 3.10 yet. Now, we upgraded
our build system to use the modern Python packaging (with
pyproject.toml), using the new editable install, and for some reason
with Python 3.10 that prevents our use of the (deprecated) `imp` builtin
module.

Since `imp` must be replaced by `importlib` anyway, and since we require
Python 3.7 minimum, it can be safely done. The code is even simpler than
before, so it's a win eitherway.