pytest: ignore pytest.PytestAssertRewriteWarning
Because sopel exposes a pytest plugin that we used with an editable
install, it looks like there is a conflict in Pytest: it tries to
rewrite the assert from the pytest_plugin.py file, even tho it's already
loaded.

As a result, it generates a warning: PytestAssertRewriteWarning.
This warning only means that the asserts in the file are not going to be
modified, which is fine by us, as this doesn't affect third-party
plugins that would use our `plugin.example` decorator.

However, the warning is annoying, so we suppress it through our test
configuration.