plugins: support multiple `sopel.modules` submodule search locations
I'll be the first to admit, this doesn't actually *change* the built-in
plugins that are enumerated. But that was also the point: It *should*
give the same list as before.

Doing this is all about code hygiene. I just don't like blindly pulling
the `[0]`th element of a list and ignoring the rest, even if that list
only has a single entry. So the code to `find_internal_plugins()` now
accounts for the possibility that there could be multiple places it has
to search for those, and deduplicates the names found before trying to
load them.

Co-authored-by: Exirel <florian.strzelecki@gmail.com>