plugin: add the capability decorator
The `capability` decorator, and the `CapabilityNegotiation` enum allow a
plugin to request capabilities.

This will replace the current system with the CapReq class and the
irc.AbstractBot.cap_req method that nobody uses because it's way too
complex and convoluted (also it doesn't do what it says it does, or not
properly).

Collecting capability requests and handling them will be part of another
commit.

Co-authored-by: dgw <dgw@technobabbl.es>