ci: select an available version of Python on Netlify's Focal image
Xenial images are being discontinued in a few months, so the time has
come for us to use Focal for the unstable docs.

Focal allegedly ships with Python 3.8 as default, but I don't trust the
documentation over at Netlify on account of having too many cases with
it and other tools that claim to default to py3 but don't actually.