remind: skip `remind_monitoring()` if IRC connection not ready
Past-due reminders could be consumed during startup (and emit an error
to the log about the EventLoop not being started yet).

The documentation for `@plugin.interval()` has had a warning in it about
not assuming that the IRC connection will be available for I don't even
know how long, at this point. It's about time core followed its advice.