trigger: first crack at fixing the inconsistent "sender" property
Short sleep does not make me a great comprehensive reader of specs, but
commands that are 1) sent from the server to a client and 2) include a
<target> (channel/nick), are actually quite rare. The IRC protocol is
mostly numeric replies, and very few event types actually refer to a
user (nick) or channel.

The idea here, though, is that unless I forgot about an event type that
actually is tied to a channel or query (PM) buffer, it is ONLY the new
list of `COMMANDS_WITH_CONTEXT` that should have the `sender` set to any
value other than `None`.

Haven't yet addressed the fact that `sender` is not the clearest name
for this property, though. Baby steps.

While updating/adding tests I also found an actual bug in how `text`
gets set (it contradicts Sopel's documentation). TODO comments related.