trigger: fix `text` to be empty if `args` is empty, like docs imply
An IRC command/event such as 'AWAY' with no message, or 'PING', will
have an empty `args` list. Documentation for the (Pre)Trigger `text`
attribute says that it will contain the last item in `args` if the line
does not contain ':', but it previously contained the command name in
such cases.