build: pin setuptools while we figure out PEP0660 editable install
In setuptools 64.0, editable installs were changed. As of 65.x, I was
unable to install sopel as an editable distribution without getting
ImportErrors even when trying to do something as simple as run `pytest
--version` (which shouldn't load any Sopel code).

I *believe* this is enough to work around the issue for now, given that
pyproject.toml packages are supposed to use build isolation and the used
setuptools version is separate from whatever's installed in the user's
environment. At least my brief testing bore this out, with ~=63.0
requested by the TOML file and 65.something installed in pip.