bot, irc, coretasks: more robust `bot.connection_registered` value
The `connection_registered` attribute itself is now a property, which
first checks to see if the bot's `backend` exists (is not `None`) and
then whether the `backend.is_connected()`, before finally seeing if the
internal `_connection_registered` flag has been set.

Also added a few places where the internal flag should be reset back to
its initial value of `False`, such as if the bot is about to quit or the
backend's `on_close()` handler was called.

All of this is to try and make `bot.connection_registered` reliable for
plugins like `remind`, which have scheduled tasks that might need to
send data to IRC and need a reliable way to check whether the connection
is ready to accept such data before operating.