tools.time: `versionadded` annotations for great justice
Was working on something in sopel-github and got annoyed that I had to
check `git blame` to verify that the functions I wanted to use would be
available in that plugin's minimum required Sopel version.

If people are nice to me, I might go through the rest of `sopel.tools`
and add any missing version annotations in its other submodules, too.