README: remove soon-to-be-defunct LGTM badge
I cannot find a way to get this same badge type with CodeQL running on
GitHub Actions. It provides a pass/fail badge for the workflow, but no
statistics views that I can find.

We didn't really use the badge for anything practical (want to know
about alerts? check the LGTM dashboard), but it's a nostalgic loss.