irc: annotate submodules with internal-use warnings
The top-level docs page for `sopel.irc` did, in fact, have an inline
note about the `irc` module being for internal use. That wouldn't catch
the eye of someone landing on a submodule via search (internal or web),
though. Seemed prudent to beef up the presentation to `.. warning::` and
also add a notice to each submodule.

Annotating `sopel.irc.isupport` was extra fun, because part of it *is*
included in the bot's public API.