tools.calculation, docs: version annotations & missing docstring
Sopelunked into the history and found when all this stuff was originally
added to the codebase. (Spoiler: A long time ago.)

There was no docstring at all for `EquationEvaluator`, so I wrote one.

Tweaked the docs page for `tools.calculation` to show everything instead
of only what's listed in `__all__`, and noted that most of the module is
considered internal machinery as we've been doing in `irc` etc.

NOTE: I would like the "public function" to appear at the top, but could
not find a nice way to do that in ~15 min of googling+experimenting.