pyproject.toml: revert "dev" dependencies
There are two kinds of "dev" dependencies:

* the ones required to develop the software
* the ones required to develop for the software

In the pyproject.toml, "dev" (or "extra") dependencies should reflect
the dependencies needed to developr for the software, such as plugins,
extensions, tools, etc.

What was here was dependencies to develop the software itself, i.e. for
sopel rockstars and contributors; and it wasn't the right place, so I
removed them.

When working on Sopel itself, we need to update frequently these
dependencies, and sadly `pip install -U -e .[dev]` didn't really gave
the expected results.

So the new workflow is to do the same thing as before:

* `pip install -e .` to install Sopel locally with an editable install
* `pip install -U -r requirements.txt` to install and update all the dev
  dependencies

Note: this helped me discover small mypy notes & errors to fix.