ip: simplify if/elif to select GeoIP database
In a previous version, Sopel was not guarantee to have the retrieve
function, that may come from different imports based on the Python
version.

Now that everything is running with Python 3+, the function always
exists, so `elif retrieve:` was always true, becoming an irrelevant
condition.

`mypy` complained, rightfully so, about it.

Co-authored-by: dgw <dgw@technobabbl.es>