coretasks: send WHO to get realname for never-seen users on JOIN
The WHO command can take a nickname, a channel name, or a pattern mask.
Sopel previously made use of only the channel name type.

I refactored the `_send_who()` helper a bit so it can take either a
channel or nick, and tweaked the JOIN handler to send a WHO for any user
Sopel didn't already know about (wasn't present in `bot.users` at the
time of the JOIN event).

This should make it so that, even on networks with `away-notify` support
(where Sopel disables its regular WHO polling), users who were not in a
given channel when Sopel joined will still have a `realname` value set
in relatively short order.