bot: don't let per-channel config skip coretasks handlers
This seems like a pretty obvious gotcha, but I'll bet most people don't
take a list of *every available plugin* and blindly paste it into the
`disable_plugins` setting for a channel.

That said, if `coretasks` is present in the `core.exclude` list, it'll
be loaded anyway. This is the same idea; Sopel NEEDS it to work, so it
shouldn't be possible to disable.

I'm not 100% sure that I needed to add checks in both of these functions
but it felt OK to be extra careful.