irc: use `threading.Event` for internal `_connection_registered` flag
This will let us add API features that take advantage of the `.wait()`
method available for Event objects.

Shouldn't affect plugin code (internal or third-party) much, if at all,
since this is an internal value. `connection_registered` (no underscore)
is still the name in all released versions of Sopel, anyway. In 8.0 it
has become a property that checks multiple pieces of state so the answer
returned can be more consistent.