py, build: drop `py` plugin from core & require external package
End of support for Python 2.7 on Google App Engine, in January 2024,
approaches. Dropping this plugin from core now will let us make the
transition to a new backend easier, without having to fit it into
Sopel's release cadence.

Marked the requirement with a TODO for 8.1, when it should become an
optional manually-installed plugin like the others that have been
packageized (ipython, spellcheck, etc.).