pronouns: pronoun.is has been dead for months; spun up our own backend
First we waited on an API that might never exist:
https://github.com/witch-house/pronoun.is/pull/96

Then the pronoun.is site went offline completely:
https://github.com/witch-house/pronoun.is/issues/147

Eventually someone came along to make a clone that was easier to host,
written using Svelte instead of Clojure:
https://github.com/lucasew/svelte-pronounisland

We can fork that so we have control over when it's updated:
https://github.com/sopel-irc/pronoun-service

And now we have our own pronoun site to go with this plugin:
https://pronouns.sopel.chat/

(But it still doesn't have an API.)

It's a shame to lose pronoun.is, though. That's the Best™ domain. :(