pronouns: configurable base URL
Since we maintain a fork of svelte-pronounisland, it's not that big of a
jump for bot owners to want a vanity instance tied to their communities.
This is a simple configuration option to offer.

It would be equally simple (but more annoying) to have a second setting
that allows overriding the URL for the `.tab` file containing all known
pronoun sets, used when `fetch_complete_list` is True. Won't cause any
problems unless we make frequent updates to our fork that aren't picked
up automatically by downstream users.