bot: eliminate problematic xrefs from `SopelWrapper` docs
The formatting, in particular, is what was problematic here. Sphinx
doesn't provide a way to negate the monospace formatting of a :meth:
reference when using a custom title, so the only way to avoid having
random monospace words in the middle of this paragraph is... rewriting
to eliminate the xrefs.

Which is to say, if there's a better answer to this question
(https://stackoverflow.com/q/45741326/5991) in the future, we can put
this back the way it was.