sopel, docs, test: replace SopelWrapper with contextvars
For ease of use, Sopel provides a SopelWrapper to plugin callable so
they can use ``bot.say(message)`` without having to specify the
destination: the wrapper will look at the trigger's sender and send the
message back to it (be it a channel or a nick).

However, this has the disavantage of having to 1. maintain a separate
class with dunder methods (boo, magic!), and 2. it makes static typing
more complex.

Welcome Python 3.7 and its new built-in module: contextvars. This
wonderful utility module provides a thread-safe and asyncio-safe object
that can have a context specific value.

Sopel now uses this feature to bind itself to:

* a trigger that matches a rule
* and the triggered rule

so now it can be used instead of a SopelWrapper instance.

The biggest changes are on ``call_rule``, which doesn't require anymore
a SopelWrapper instance, and on the tests, in particular on the trigger
factory's wrapped method, which now returns a Sopel instance.

The second side-effect on test is that now it is not possible to bind a
trigger multiple times on a test bot, and tests have to rely on
Sopel.sopel_wrapper context manager (as Sopel.call_rule does), which
requires a Rule object. This is not perfect, albeit manageable.

In the future, it would be nice to have a better testing utility, maybe
a context manager to do someething like:

```
def some_test(mockbot, triggerfactory):
    with triggerfactory.wraps(mockbot, "PRIVMSG #channel Nick :text") as wrapped:
        wrapped.say('something')
```

Wouldn't it be cool?