reddit: extract to external package
The writing's on the wall:
https://old.reddit.com/r/reddit/comments/12qwagm/an_update_regarding_reddits_api/

It's reasonable to guess that these changes to rate-limit enforcement
(however vague they might be now) will eventually affect Sopel bots
running with the default, shared, OAuth app ID.

A plugin that no longer works (reliably) without the bot owner having to
acquire an API key of some kind fails our number-one criterion for
inclusion in the base Sopel package. While the reddit plugin hasn't yet
reached that point, we are also trying to extract a lot of the built-ins
anyway for maintenance reasons.

That transition is much easier to make on our own schedule, right before
a major release. And a major release is a doubly great time for this
particular change, because we can drop `praw` as a core requirement.