tools.target: finish type-hinting module
Converts `User.hostmask` from a lambda to a proper function, in order to
apply the intended type hints. (Setting `hostmask: str = property(...)`
generated an error, of course. It was also not possible to type-hint the
lambda's return value without stooping to `typing.Callable[]`, and I
figured I might as well just make the thing a real function.)

TODO comment added regarding user/host values that haven't populated yet
and are still set to the default of `None`.