build: pin urllib3 to <2.0 for now
CI is broken with urllib3 2.x, but the specific package that fails our
tests (vcrpy) is not moving very swiftly to release a fix.

While they debate which versions of urllib3 to support, we just want to
have working CI so we can start merging our slowly growing PR backlog.

Because urllib3 is a transitive dependency of requests, it belongs in
the regular requirement list even though this pin is to fix tests.