cli.plugins: tweaks to fix typing errors
mypy does not like assigning a list to an existing variable that was
already assigned as a generator. This is fair. It'd likely be equally
happy with annotating `items` as an Iterable, but I'm not type-hinting
whole files in this housekeeping pass.

Likewise, later in the file, mypy does not like assigning a tuple to a
variable that was declared initially as an empty list. That's also fair,
but it also didn't like replacing the initial `actually_disabled = []`
with `actually_disabled = tuple()`, so I simply removed it. There's no
risk of the variable being undefined, as the comprehension expression
always runs (even if it yields zero items).