cli.run: 'action' option can never be None, so why default to that?
mypy gets upset at subscripting a dict using an Optional[str] key, and
rightly so. The following code above the option-parsing line always adds
the 'start' action to the command line if there is no other action set:

    elif argv[0].startswith('-') and argv[0] not in global_options:
        # No sub-command and no global option
        argv = ['start'] + argv

The parser will always have an 'action' to pass in, therefore we don't
need the third argument to set `None` as the `getattr()` default.