plugin: type-hint `example` class + fix minor typing issue
mypy needs to know the type of `example.result`, and it wouldn't have
made much sense to add only that without also telling mypy what the
function argument being assigned to the instance attribute was... and
it wouldn't make sense to type-hint only one argument to the
constructor method.

So here we are, with type-hints for ALL the constructor parameters, and
improved logic for dealing with `result` & `ignore` (both of which take
string OR an iterable OF strings).

Wanted to type-hint `example.__call__()` too, but that'll require some
help from custom types defined elsewhere (that don't yet exist).