bot: don't add duplicate timestamp into error log line
If a Trigger causes an error in a plugin, we already know when the error
happened because the logger call is timestamped. The error as reported
to IRC also is timestamped, and `datetime.now()` is not useful data.

(It is, however, confusing data. Considered replacing it with
`trigger.time` but decided against that, since `server-time` means those
could be in the past/future, making the timestamp embedded in the log
message even *more* confusing instead of less.)