plugins.rules: clean up forgotten py2 list-copy holdover
Not doing exactly as the comment says because different Rule subclasses
pass aliases as either tuple or list depending on how whoever wrote the
subclass was feeling that day.

Instead, let's simplify the code to do everything in one step, and
unpack the `aliases` iterable so its type won't even matter.

Co-authored-by: SnoopJ <snoopjedi@gmail.com>