test: fix incorrect capability NAK test names
It's not much, but it's honest work.

I was going to do some type-hint cleanup too, but there are plenty of
warnings in the `test/` directory because we haven't been checking it.
That sort of work should be done in its own PR(s) rather than as
incidental cleanup "because I'm already touching this file".