tools.memories, test: handle creating `SopelIdentifierMemory` from seqs
I'm timid about changing the actual method signature for `__init__()`,
but could be talked into letting Python itself raise the `TypeError` for
us if this approach looks too dumb. (All of the memory types can be
cleaned up at once later, which is another reason I chose not to change
the signature for just one of them.)

But here we are: `SopelIdentifierMemory.__init__()` now handles being
passed an iterable of `(key, value)` 2-tuples (whether the iterable
itself is a list or a tuple), a regular `dict`, and an existing
`SopelIdentifierMemory` object.

In that last case, it's doing some unnecessary work in most cases by
always casting the keys to whatever the `identifier_factory` function
returns, but there are *probably* a few obscure cases where that might
actually be useful behavior so I'll leave it.