test: verify `SopelIdentifierMemory.clear()` behavior
Prior to some of the recent work done on this class, calling `.clear()`
on a `SopelIdentifierMemory` instance would turn that object into a
normal `dict`, according to my tests a little while ago.

That appears to no longer be the case, and though I didn't explicitly
work on that problem yet, a unit test to make sure the behavior isn't
broken again will do us good.