doc, test: clarify that `SopelMemory` &c. can't be init-ed w/kwargs
Constructing a `SopelMemory` by passing kwargs as key-value pairs is not
currently supported. This applies to all derived types, and since there
are now a bunch of tests for `SopelIdentifierMemory`, it gets to be the
guinea pig for an xfail-marked test of that fact.

The same limitation applies to `SopelMemoryWithDefault`, since it also
shares `SopelMemory`'s constructor method signature (which does not
accept any kwargs).

Co-authored-by: Exirel <florian.strzelecki@gmail.com>