dice: remove dead code for `addition` argument to `DicePouch`
The equation evaluator has been taking care of this since 2013, when the
`addition` argument was hard-coded to `0` in a dice-rolling rewrite. See
commit 5a8ecf402bc79976f74c39b7ff4b558b65fe099d for some of the history.

A decade later, in 2023, we have no use for this argument. The code to
handle it is unreachable, which affects the plugin's coverage score.
Best solution is to just remove the dead code.