dice: add `plugin.example` tests for most missed branches/lines
There are still 3 missed lines (2 partial branches) that I don't believe
are possible to test using the example decorator. Both missed branches
check for invalid input: one for a non-string expression hypothetically
passed to `eval_equation()`, and the other for a non-dice-like pattern
hypothetically passed to `_roll_dice()`. So much input manipulation is
already happening upstream of these sanity checks that it's *probably*
impossible to reach them.

For now, I'll be satisfied with 98% coverage in this file. In a later
development cycle (i.e. not for 8.0), maybe we should convert this set
of tests to a regular test file in `test/modules/test_modules_calc.py`
and leave the examples for user documentation only. That would let us
have actual *unit* tests for the various components, and cover the
probably-unreachable-but-makes-us-feel-safe sanity checks too.