dice: type hints & docstring tweaks
Not that these docstrings are visible anywhere, but it still feels good.

One particularly tricky thing with mypy is because `_roll_dice()` is
allowed to return `None`, so when building a list of dice rolls the type
of the resulting variable MUST be `list[Optional[DicePouch]]`. This is a
problem later, where the values of that list are passed to private
helper functions that only accept `DicePouch`, not `Optional[DicePouch]`
(i.e. they don't handle `None`).

My solution was to save dice-roll results as `list[Optional[DicePouch]]`
and check if any of them returned `None` to indicate a problem (aborting
 if so), then filter the results into a new `list[DicePouch]` variable
for further processing. I don't like it that much, but it was the lowest
impact change I saw available. The "more correct" solution is probably
refactoring to make `DicePouch` raise an exception on error, instead of
returning `None`, but that seemed out of scope for an annotations patch.