docs: make autodoc preserve default arguments without evaluating them
This is a supposedly "experimental" option that was added in Sphinx 4.0,
but here we are using Sphinx 7.1.x and it still hasn't been added to
autodoc core as planned.

Turning it on fixes the display of default parameters values like
`tools.memories._NO_DEFAULT` in function/method signatures.

Admittedly, I have not checked through the entire documentation set
again after enabling this, but it seems unlikely that there are ever
cases where we want the evaluated value to display instead of how it's
passed in the source code.