docs: refer only to the current commit
The output of `git-describe` is in terms of the latest tag on the
current branch, which can produce surprising results when tags are
distributed across release branches (i.e. `v7.1.0-992-g51300a1a` is the
best it can produce for that commit, even though the latest stable tag
at that point is `v7.1.9`).

It makes more sense to refer only to the commit.