plugins.rules: fix warnings from Sphinx unable to find `datetime` class
We have the choice of two not-great situations:

1. Concise `datetime` imports with broken cross-references in docs.
2. Verbose `datetime` imports with working cross-references in docs.

The ideal, of course, would be "Concise `datetime` imports with working
cross-references in docs", but this seems to be an impossibility at the
current time. Forced to choose between less-than-ideal alternatives, I
say we should pick the one that leads to better documentation. The cost
is cheap, at just 9 extra characters in each place where `plugins.rules`
references a type from sdlib's `datetime` module.