docs: clarify logging config overview
There is always a log format and a log date format because those options
in the core section have defaults. Likewise, the default logging level
of INFO exists even if not set explicitly by the config file.

(Removed the confusing bit about WARNING being the default; that's the
default of the `level` argument to `logging.basicConfig()`, but Sopel
always passes the value of its `logging_level` setting.)