docs: add note about scheduling function calls to `plugin/advanced`
We learned why replicating the API of another library sucks back when
`sopel.web` contained wrappers for various tools from `requests`, and we
don't want to do it again.

Setting up a scheduler for one's plugin doesn't take a lot of work. It's
not ideal for multiple plugins to all have their own `interval` checks
for whether there are any pending tasks to run, but we can drop this new
section from the docs if we find a decent compromise.