test: verify more `find` behaviors
Check that correcting someone else's line works as expected, and check
that replacing an already-replaced line works as expected.

Also check that case-insensitive (i) and global (g) flags work as
expected, both separately and together.

Note: There is a glitch in formatting when the new replacement occurs
inside the previous one, since the bolding used to indicate what changed
is kept in the stored line. We might wish to fix that at some point, but
the new test correctly represents the real behavior.