find: fix bolding when replacing inside a previous replacement
Perform the replacement with an unformatted version of the substitute
string, then create a display version with bolding added only after the
matching line (if any) is found.

Updated the corresponding test case. Also renamed some local vars in the
`find` plugin file to make a bit more sense (but not too dramatic).